Restructuring admin a little, making description and biography textareas
