+++ src/main/webapp/app/controls/admin/admin-tabs/admin-organizations/admin-organizations-form.component.ts
@@ -2,15 +2,15 @@ import {Component, Input, OnInit} from "@angular/core";
 import {FormBuilder, FormGroup, Validators} from "@angular/forms";
 import {MdDialogRef} from "@angular/material";
 
-import {AdminDialogComponent} from "../../../admin-dialog.component";
-import {AdminModel} from "../../../admin.constants";
-import {AdminService} from "../../../../../services/admin.service";
-import {NotifyService} from "../../../../../services/notify.service";
+import {AdminDialogComponent} from "../../admin-dialog.component";
+import {AdminModel} from "../../admin.constants";
+import {AdminService} from "../../../../services/admin.service";
+import {NotifyService} from "../../../../services/notify.service";
 
 @Component({
   selector: 'admin-organizations-form',
   templateUrl: './admin-organizations-form.component.html',
-  styleUrls: ['../../admin-forms.css']
+  styleUrls: ['../admin-forms.css']
 })
 export class AdminOrganizationsFormComponent implements OnInit {
