+++ src/main/webapp/app/controls/admin/admin-tabs/admin-sessions/admin-sessions-form.component.ts
@@ -2,15 +2,15 @@ import {Component, Input, OnInit} from "@angular/core";
 import {FormBuilder, FormGroup, Validators} from "@angular/forms";
 import {MdDialogRef} from "@angular/material";
 
-import {AdminDialogComponent} from "../../../admin-dialog.component";
-import {AdminModel} from "../../../admin.constants";
-import {AdminService} from "../../../../../services/admin.service";
-import {NotifyService} from "../../../../../services/notify.service";
+import {AdminDialogComponent} from "../../admin-dialog.component";
+import {AdminModel} from "../../admin.constants";
+import {AdminService} from "../../../../services/admin.service";
+import {NotifyService} from "../../../../services/notify.service";
 
 @Component({
   selector: 'admin-sessions-form',
   templateUrl: './admin-sessions-form.component.html',
-  styleUrls: ['../../admin-forms.css']
+  styleUrls: ['../admin-forms.css']
 })
 export class AdminSessionsFormComponent implements OnInit {
